Complete working Acl example, with getList

+++ application/models/AclHandler/Entry.php
@@ -28,6 +28,60 @@ class Default_Model_AclHandler_Entry
     }
 
     /**
+     * @param array $params
+     * @return array
+     */
+    public function getList(array $params = null)
+    {
+        $username = $this->getAclContextUser();
+
+        if ($this->isAllowed('get')) {
+            // get excluding the blacklist
+            // accept specific resources that are allow or unspecified.
+            // IE: not denied
+
+            $sql = 'SELECT e.id, e.comment, e.created'
+                . ' FROM entry AS e'
+                . ' LEFT OUTER JOIN permission AS p ON p.resource = ("Entry=" || e.id)'
+                . ' LEFT OUTER JOIN role AS r ON p.role = r.role AND p.resource = r.resource'
+                . ' LEFT OUTER JOIN user AS u ON r.user_id = u.id'
+                . ' WHERE p.id IS NULL OR ('
+                . '     ('
+                . '         u.username = :username'
+                . '         OR p.role = "default"'
+                . '     )'
+                . '     AND p.privilege = "get"'
+                . '     AND p.permission != "deny"'
+                . ' )'
+                . '';
+        } else {
+            // get based on whitelist
+            // accept specific resource that are allow only
+
+            $sql = 'SELECT e.id, e.comment, e.created'
+                . ' FROM entry AS e'
+                . ' LEFT OUTER JOIN permission AS p ON p.resource = ("Entry=" || e.id)'
+                . ' LEFT OUTER JOIN role AS r ON p.role = r.role AND p.resource = r.resource'
+                . ' LEFT OUTER JOIN user AS u ON r.user_id = u.id'
+                . ' WHERE ('
+                . '     u.username = :username'
+                . '     OR p.role = "default"'
+                . ' )'
+                . ' AND p.privilege = "get"'
+                . ' AND p.permission = "allow"'
+                . '';
+        }
+
+        $query = $this->_getDbHandler()->prepare($sql);
+        $query->execute(array(
+            ':username' => $username,
+        ));
+        $rowSet = $query->fetchAll(PDO::FETCH_ASSOC);
+
+        return $rowSet;
+    }
+
+    /**
      * @param array $id
      * @return array
      * @throws Rest_Model_NotFoundException, Zend_Acl_Exception
@@ -49,6 +103,10 @@ class Default_Model_AclHandler_Entry
      */
     public function put(array $id, array $prop = null)
     {
+        if ($this->getAcl() && !$this->isAllowed('put', $id)) {
+            throw new Zend_Acl_Exception('put for ' . $this->getResourceId() . ' is not allowed');
+        }
+
         return $this->_getModelHandler()->put($id, $prop);
     }
 
@@ -58,6 +116,10 @@ class Default_Model_AclHandler_Entry
      */
     public function delete(array $id)
     {
+        if ($this->getAcl() && !$this->isAllowed('delete', $id)) {
+            throw new Zend_Acl_Exception('delete for ' . $this->getResourceId() . ' is not allowed');
+        }
+
         return $this->_getModelHandler()->delete($id);
     }
 
@@ -68,17 +130,11 @@ class Default_Model_AclHandler_Entry
      */
     public function post(array $prop)
     {
+        if ($this->getAcl() && !$this->isAllowed('post', $id)) {
+            throw new Zend_Acl_Exception('post for ' . $this->getResourceId() . ' is not allowed');
+        }
+
         return $this->_getModelHandler()->post($prop);
-    }
-
-
-    /**
-     * @param array $params
-     * @return array
-     */
-    public function getList(array $params = null)
-    {
-        return $this->_getModelHandler()->getList($params);
     }
 
     /**
